Experimental support for projection.origin.
I still think it makes more sense to have a d3.geo.rotate operator which
operates on GeoJSON features, so that features can be split when they cross the
antemeridian or the poles. Although, if projections could report their extent
(e.g., the graticule boundary in latitude and longitude), then potentially we
could do the clipping inside d3.geo.path.