Add projection.center and projection.rotate.
This clarifies the related concepts of origin and translation. The projection
center provides a convenient way of controlling the longitude and latitude that
is projected to the center of the canvas, as determined by the projection's
translation; projection.center is similar to the center method on popular
JavaScript mapping libraries. The projection rotation is applied to input
coordinates before projecting.

For backwards-compatibility and for convenience, the projection.origin method is
retained, and is equivalent to setting the rotation longitude and the center
latitude. This change is not strictly backwards-compatible because of the
interaction with the projection's translation, however I think it's appropriate
to think of this as a bug fix rather than an API breakage.