d3.simplify: improve topology preservation.
Instead of using effective areas of ∞ for shared endpoints, we now use
the maximum effective area for each connected component of the topology.
This allows islands to be removed if their maximum effective area is
below the threshold. This also fixes an issue when the projected
features are encoded as JSON, since JSON converts Infinity to null.

Efficiency should also be somewhat improved.

Related: #21.