Add auxiliary stream transformation.
This allows the spherical form of a map projection to be adapted for use
with the ellipsoid by substitution of an “auxiliary” latitude in place
of the geographic latitude.  The geographic latitude normally refers to
the angle a line perpendicular to the surface of the ellipsoid at a
particular point makes with the equatorial plane.  Auxiliary latitudes
transform the ellipsoid to a sphere under restraints such as
conformality or equal area (authalic).