Rewrite d3.geo.gilbert.
It is impossible to make this a useful d3.geo.projection instance, since
rotation needs to be applied after Gilbert’s conformal transformation on
the sphere, but before the final orthographic projection.  As
d3.geo.projection takes care of rotation for us, as well as automatic
centring, it is simpler and more useful to allow wrapping of any target
projection.